requirements.txt
asgiref==3.3.1
Django==3.1.6
django-crispy-forms==1.11.0
django-leaflet==0.27.1
django-summernote==0.8.11.6
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
